+++ version/info.go
@@ -17,7 +17,7 @@ import "fmt"
 
 // These three variables are overwritten by -ldflags at build time.
 var (
-	Version   = "v2.107.0" // semver tag  e.g. "v2.99.0"
+	Version   = "v2.108.0" // semver tag  e.g. "v2.99.0"
 	Commit    = "none"    // git SHA     e.g. "abc1234"
 	BuildDate = "unknown" // build date  e.g. "2024-06-01"
 )